Consolidated all HPOA ingest YAML and Python scripts into single HPOA (yaml/py) files. All tests pass (but need to revisit MONDO)
