--- src/translator_ingest/ingests/hpoa/disease_to_phenotype_transform.yaml
@@ -1,58 +1,0 @@
-transform_code: "src/translator_ingest/ingests/hpoa/disease_to_phenotype_transform.py"
-
-name: 'hpoa_disease_to_phenotype'
-
-metadata:
-  ingest_title: 'Human Phenotype Ontology Annotations'
-  ingest_url: 'https://hpo.jax.org/app/'
-  description: 'Ingest of HPO disease to phenotype associations'
-  rights: 'https://hpo.jax.org/app/license'
-
-files:
-  - 'data/hpoa/phenotype.hpoa'
-
-depends_on:
-  - 'data/hpoa/hp.obo'
-
-header: 4
-format: 'csv'
-delimiter: '\t'
-
-# For a csv/tsv file, list expected columns
-columns:
-  - 'database_id'
-  - 'disease_name'
-  - 'qualifier'
-  - 'hpo_id'
-  - 'reference'
-  - 'evidence'
-  - 'onset'
-  - 'frequency'
-  - 'sex'
-  - 'modifier'
-  - 'aspect'
-  - 'biocuration'
-
-# Core Biolink slots common to both the
-# 'DiseaseOrPhenotypicFeatureToModeOfGeneticInheritanceAssociation', and
-# 'DiseaseToPhenotypicFeatureAssociation' Associations
-edge_properties:
-  - 'id'
-  - 'category'
-  - 'subject'
-  - 'predicate'
-  - 'object'
-  - 'publications'
-  - 'has_evidence'
-  - 'sources'  # primary & supporting
-  - 'knowledge_level'
-  - 'agent_type'
-  # Additional annotation specific to DiseaseToPhenotypicFeatureAssociation
-  - 'negated'
-  - 'onset_qualifier'
-  - 'frequency_qualifier'
-  - 'has_count'
-  - 'has_total'
-  - 'has_percentage'
-  - 'has_quotient'
-  - 'sex_qualifier'
