--- src/translator_ingest/ingests/hpoa/gene_to_disease_transform.yaml
@@ -1,30 +1,0 @@
-transform_code: "src/translator_ingest/ingests/hpoa/gene_to_disease_transform.py"
-
-name: 'hpoa_gene_to_disease'
-
-files:
-  - './data/hpoa/genes_to_disease.txt'
-
-format: 'csv'
-
-delimiter: '\t'
-
-columns:
-  - 'ncbi_gene_id'
-  - 'gene_symbol'
-  - 'association_type'
-  - 'disease_id'
-  - 'source'
-
-edge_properties:
-  - 'id'
-  - 'category'
-  - 'subject'
-  - 'predicate'
-  - 'object'
-  - 'aggregator_knowledge_source'
-  - 'primary_knowledge_source'
-  - 'knowledge_level'
-  - 'agent_type'
-
-transform_mode: 'flat'
