--- src/translator_ingest/ingests/hpoa/gene_to_phenotype_transform.yaml
@@ -1,65 +1,0 @@
-# Config file for transforming data from a source
-# See additional/optional config parameters at https://koza.monarchinitiative.org/Ingests/source_config/
-
-transform_code: "src/monarch_phenotype_profile_ingest/gene_to_phenotype_transform.py"
-
-name: "hpoa_gene_to_phenotype" # This is the name of the koza app that is passed into the transform.py
-metadata: "src/translator_ingest/ingests/hpoa/metadata.yaml"
-
-# This file is produced by gene_to_phenotype_extras.py using genes_to_phenotype.txt,
-# publications from phenotype.hpoa, and association types from genes_to_disease.txt
-files:
-  - 'data/hpoa/genes_to_phenotype_preprocessed.tsv'
-
-header: 0
-format: 'csv'
-delimiter: '\t'
-
-columns:
-  - 'ncbi_gene_id'
-  - 'gene_symbol'
-  - 'hpo_id'
-  - 'hpo_name'
-  - 'frequency'
-  - 'disease_id'
-  - 'publications'
-  - 'gene_to_disease_association_types'
-
-edge_properties:
-  - 'id'
-  - 'category'
-  - 'subject'
-  - 'predicate'
-  - 'object'
-  - 'aggregator_knowledge_source'
-  - 'primary_knowledge_source'
-  - 'knowledge_level'
-  - 'agent_type'
-  - 'frequency_qualifier'
-  - 'has_count'
-  - 'has_total'
-  - 'has_percentage'
-  - 'has_quotient'
-  - 'disease_context_qualifier'
-  - 'publications'
-
-transform_mode: 'flat'
-
-filters:
-  - inclusion: 'include'
-    column: 'gene_to_disease_association_types'
-    filter_code: 'in'
-    value:
-      - 'MENDELIAN'
-
-# We are using the mapping table functionality instead (may switch to this down the road...)
-# # For mondo disease_id mapping
-# sssom_config:
-#   files:
-#    - "data/mondo.sssom.tsv"
-#   subject_target_prefixes:
-#    - "MONDO"
-#   use_match:
-#    - "exact"
-  
-
