+++ src/translator_ingest/ingests/hpoa/hpoa.yaml
@@ -1,0 +1,99 @@
+id:  'infores:hpo-annotations'
+name: 'Human Phenotype Ontology Annotations'
+metadata: !include './metadata.yaml'
+
+readers:
+  disease_to_phenotype:
+    transform_code: 'src/translator_ingest/ingests/hpoa/disease_to_phenotype_transform.py'
+    name: 'hpoa_disease_to_phenotype'
+    files:
+      - 'data/hpoa/phenotype.hpoa'
+    depends_on:
+      - 'data/hpoa/hp.obo'
+    header: 4
+    format: 'csv'
+    delimiter: '\t'
+    # For a csv/tsv file, list expected columns
+    columns:
+      - 'database_id'
+      - 'disease_name'
+      - 'qualifier'
+      - 'hpo_id'
+      - 'reference'
+      - 'evidence'
+      - 'onset'
+      - 'frequency'
+      - 'sex'
+      - 'modifier'
+      - 'aspect'
+      - 'biocuration'
+
+  gene_to_disease:
+    transform_code: "src/translator_ingest/ingests/hpoa/gene_to_disease_transform.py"
+    name: 'hpoa_gene_to_disease'
+    files:
+      - './data/hpoa/genes_to_disease.txt'
+    format: 'csv'
+    delimiter: '\t'
+    columns:
+      - 'ncbi_gene_id'
+      - 'gene_symbol'
+      - 'association_type'
+      - 'disease_id'
+      - 'source'
+    transform_mode: 'flat'
+
+  gene_to_phenotype:
+    transform_code: "src/monarch_phenotype_profile_ingest/gene_to_phenotype_transform.py"
+
+    name: "hpoa_gene_to_phenotype" # This is the name of the koza app that is passed into the transform.py
+    metadata: "src/translator_ingest/ingests/hpoa/metadata.yaml"
+
+    # This file is produced by gene_to_phenotype_extras.py using genes_to_phenotype.txt,
+    # publications from phenotype.hpoa, and association types from genes_to_disease.txt
+    files:
+      - 'data/hpoa/genes_to_phenotype_preprocessed.tsv'
+
+    header: 0
+    format: 'csv'
+    delimiter: '\t'
+
+    columns:
+      - 'ncbi_gene_id'
+      - 'gene_symbol'
+      - 'hpo_id'
+      - 'hpo_name'
+      - 'frequency'
+      - 'disease_id'
+      - 'publications'
+      - 'gene_to_disease_association_types'
+    transform_mode: 'flat'
+    filters:
+      - inclusion: 'include'
+        column: 'gene_to_disease_association_types'
+        filter_code: 'in'
+        value:
+          - 'MENDELIAN'
+
+  mondo_map:
+    name: 'mondo_map'
+    files:
+      - './data/mondo.sssom.tsv'
+    format: 'csv'
+    delimiter: '\t'
+    header: 51
+    filters:
+      - inclusion: 'include'
+        column: 'predicate_id'
+        filter_code: 'eq'
+        value: 'skos:exactMatch'
+    key: object_id
+    values:
+      - subject_id
+    columns:
+      - subject_id
+      - subject_label
+      - predicate_id
+      - object_id
+      - object_label
+      - mapping_justification
