--- src/translator_ingest/ingests/hpoa/mondo_sssom_config.yaml
@@ -1,27 +1,0 @@
-name: 'mondo_map'
-
-files:
-  - './data/mondo.sssom.tsv'
-
-format: 'csv'
-delimiter: '\t'
-header: 51
-
-filters:
-  - inclusion: 'include'
-    column: 'predicate_id'
-    filter_code: 'eq'
-    value: 'skos:exactMatch'
-
-key: object_id
-
-values:
-  - subject_id
-
-columns:
-  - subject_id
-  - subject_label
-  - predicate_id
-  - object_id
-  - object_label
-  - mapping_justification
